server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
diffusers>=0.21.0
torch>=2.0.0
Pillow>=10.0.0
accelerate>=0.20.0
transformers>=4.30.0


